Fix PIO warnings and rename test directories so they work!
