+++ src/platformio.ini
@@ -36,7 +36,7 @@ platform = native
 framework =
 test_ignore = test_embedded
 lib_ignore = BUTTON, DAC, LQCALC
-src_filter = ${common_env_data.src_filter} -<ESP32*.*> -<STM32*.*> -<ESP8*.*> -<tx_*.cpp> -<rx_*.cpp> -<common.*> -<config.*>
+build_src_filter = ${common_env_data.build_src_filter} -<ESP32*.*> -<STM32*.*> -<ESP8*.*> -<tx_*.cpp> -<rx_*.cpp> -<common.*> -<config.*>
 build_flags =
 	-std=c++11
 	-Iinclude
