+++ src/targets/Jumper_2400.ini
@@ -13,7 +13,7 @@ build_flags =
 	-include target/Jumper_AION_2400_T-Pro_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 upload_speed = 460800
 
 [env:Jumper_AION_2400_T-Pro_TX_via_UART]
@@ -32,7 +32,7 @@ build_flags =
 	-include target/Jumper_AION_NANO_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
 	${env_common_esp32.oled_lib_deps}
