+++ src/targets/MATEK_2400.ini
@@ -11,7 +11,7 @@ build_flags =
 	${common_env_data.build_flags_rx}
 	-include target/MATEK_2400_RX.h
 	-D USE_DIVERSITY=1
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:MATEK_2400_RX_R24D_via_BetaflightPassthrough]
 extends = env:MATEK_2400_RX_R24D_via_UART
@@ -31,7 +31,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/MATEK_2400_RX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:MATEK_2400_RX_via_BetaflightPassthrough]
 extends = env:MATEK_2400_RX_via_UART
