+++ src/targets/axis_2400.ini
@@ -14,7 +14,7 @@ build_flags =
 	-D VTABLES_IN_FLASH=1
 	-O2
 monitor_speed = 420000
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.tft_lib_deps}
 
@@ -32,7 +32,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/AXIS_THOR_2400_RX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:AXIS_THOR_2400_RX_via_WIFI]
 extends = env:AXIS_THOR_2400_RX_via_UART
