+++ src/targets/betafpv_2400.ini
@@ -12,7 +12,7 @@ build_flags =
 	-D VTABLES_IN_FLASH=1
 	-O2
 upload_speed = 460800
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:BETAFPV_2400_TX_via_WIFI]
 extends = env:BETAFPV_2400_TX_via_UART
@@ -27,7 +27,7 @@ build_flags =
 	-D VTABLES_IN_FLASH=1
 	-O2
 upload_speed = 460800
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
 
@@ -57,7 +57,7 @@ build_flags =
 	-include target/BETAFPV_2400_TX_LITERADIO3.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 upload_speed = 460800
 lib_deps =
 	${env_common_esp32.lib_deps}
@@ -80,7 +80,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/BETAFPV_2400_RX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:BETAFPV_Nano_2400_RX_via_BetaflightPassthrough]
 extends = env:BETAFPV_Nano_2400_RX_via_UART
