+++ src/targets/betafpv_900.ini
@@ -10,7 +10,7 @@ build_flags =
 	${common_env_data.build_flags_tx}
 	-include target/BETAFPV_900_TX.h
 upload_speed = 460800
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:BETAFPV_900_TX_via_WIFI]
 extends = env:BETAFPV_900_TX_via_UART
@@ -23,7 +23,7 @@ build_flags =
 	${common_env_data.build_flags_tx}
 	-include target/BETAFPV_900_TX_MICRO.h
 upload_speed = 460800
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
 
@@ -41,7 +41,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/DIY_900_RX_ESP8285_SX127x.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:BETAFPV_900_RX_via_BetaflightPassthrough]
 extends = env:BETAFPV_900_RX_via_UART
