+++ src/targets/common.ini
@@ -11,7 +11,7 @@ monitor_dtr = 0
 monitor_rts = 0
 
 [common_env_data]
-src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<test/> -<tests/> -<*.py> -<*test*.*>
+build_src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<test/> -<tests/> -<*.py> -<*test*.*>
 build_flags = -Wall -Iinclude
 build_flags_tx = -DTARGET_TX=1 ${common_env_data.build_flags}
 build_flags_rx = -DTARGET_RX=1 ${common_env_data.build_flags}
@@ -42,7 +42,7 @@ build_flags =
 	-D BEARSSL_SSL_BASIC
 	-D NO_GLOBAL_SERIAL
 	-I ${PROJECTSRC_DIR}/hal
-src_filter = ${common_env_data.src_filter} -<ESP8266*.*> -<STM32*.*>
+build_src_filter = ${common_env_data.build_src_filter} -<ESP8266*.*> -<STM32*.*>
 lib_deps =
 	makuna/NeoPixelBus @ 2.6.9
 	ottowinter/ESPAsyncWebServer-esphome @ 2.1.0
@@ -70,7 +70,7 @@ build_flags =
 	-I ${PROJECTSRC_DIR}/hal
 board_build.f_cpu = 160000000L
 board_build.ldscript = eagle.flash.1m.ld
-src_filter = ${common_env_data.src_filter} -<ESP32*.*> -<STM32*.*> -<WS281B*.*>
+build_src_filter = ${common_env_data.build_src_filter} -<ESP32*.*> -<STM32*.*> -<WS281B*.*>
 extra_scripts =
 	${env.extra_scripts}
 	pre:python/build_html.py
@@ -97,7 +97,7 @@ build_flags =
 	-I ${PROJECTSRC_DIR}/hal
 	-D __FILE__='""'
 	-Wno-builtin-macro-redefined
-src_filter = ${common_env_data.src_filter} -<ESP32*.*> -<ESP8266*.*> -<WS281B*.*>
+build_src_filter = ${common_env_data.build_src_filter} -<ESP32*.*> -<ESP8266*.*> -<WS281B*.*>
 lib_deps =
 	paolop74/extEEPROM @ 3.4.1
 oled_lib_deps =
