+++ src/targets/diy_2400.ini
@@ -12,7 +12,7 @@ build_flags =
 	-include target/DIY_2400_TX_ESP32_SX1280_Mini.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:DIY_2400_TX_ESP32_SX1280_Mini_via_WIFI]
 extends = env:DIY_2400_TX_ESP32_SX1280_Mini_via_UART
@@ -26,7 +26,7 @@ build_flags =
 	-include target/DIY_2400_TX_ESP32_SX1280_E28.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
 
@@ -42,7 +42,7 @@ build_flags =
 	-include target/DIY_2400_TX_ESP32_SX1280_LORA1280F27.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
 
@@ -58,7 +58,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_tx}
 	-include target/DIY_2400_TX_ESP8285_SX1280.h
-src_filter = ${env_common_esp82xx.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<rx_*.cpp>
 upload_speed = 921600
 
 [env:DIY_2400_TX_DUPLETX_via_ETX]
@@ -70,7 +70,7 @@ build_flags =
 	-include target/DIY_2400_TX_DUPLETX_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 upload_speed = 460800
 
 [env:DIY_2400_TX_DUPLETX_via_UART]
@@ -91,7 +91,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/DIY_2400_RX_ESP8285_SX1280.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:DIY_2400_RX_ESP8285_SX1280_via_BetaflightPassthrough]
 extends = env:DIY_2400_RX_ESP8285_SX1280_via_UART
@@ -109,7 +109,7 @@ build_flags =
 	${common_env_data.build_flags_rx}
 	${radio_2400.build_flags}
 	-include target/DIY_2400_RX_VTX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:DIY_2400_RX_VTX_via_BetaflightPassthrough]
 extends = env:DIY_2400_RX_VTX_via_UART
@@ -127,7 +127,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/DIY_2400_RX_PWMP.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:DIY_2400_RX_PWMP_via_WIFI]
 extends = env:DIY_2400_RX_PWMP_via_UART
@@ -161,7 +161,7 @@ build_flags =
 	-D VECT_TAB_OFFSET=0x4000U
     -D FLASH_APP_OFFSET=0x4000U
     -Wl,--defsym=FLASH_APP_OFFSET=16K
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/sx1280_rx_nano_pcb_v0.5_bootloader.bin
     VECT_OFFSET=0x4000
