+++ src/targets/diy_900.ini
@@ -10,7 +10,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_tx}
 	-include target/DIY_900_TX_TTGO_V1_SX127x.h
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
 
@@ -24,7 +24,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_tx}
 	-include target/DIY_900_TX_TTGO_V2_SX127x.h
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
 
@@ -38,7 +38,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_tx}
 	-include target/DIY_900_TX_ESP32_SX127x_E19.h
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:DIY_900_TX_ESP32_SX127x_E19_via_WIFI]
 extends = env:DIY_900_TX_ESP32_SX127x_E19_via_UART
@@ -50,7 +50,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_tx}
 	-include target/DIY_900_TX_ESP32_SX127x_RFM95.h
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:DIY_900_TX_ESP32_SX127x_RFM95_via_WIFI]
 extends = env:DIY_900_TX_ESP32_SX127x_RFM95_via_UART
@@ -69,7 +69,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/DIY_900_RX_ESP8285_SX127x.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:DIY_900_RX_ESP8285_SX127x_via_BetaflightPassthrough]
 extends = env:DIY_900_RX_ESP8285_SX127x_via_UART
@@ -87,7 +87,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/DIY_900_RX_HUZZAH_RFM95W.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:DIY_900_RX_HUZZAH_RFM95W_via_BetaflightPassthrough]
 extends = env:DIY_900_RX_HUZZAH_RFM95W_via_UART
@@ -105,7 +105,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/DIY_900_RX_PWMP.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:DIY_900_RX_PWMP_via_WIFI]
 extends = env:DIY_900_RX_PWMP_via_UART
