+++ src/targets/diy_ble_joystick.ini
@@ -13,6 +13,6 @@ build_flags =
 	-include target/DIY_2400_TX_ESP32_SX1280_E28.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
     ${env_common_esp32.oled_lib_deps}
