+++ src/targets/frsky.ini
@@ -16,7 +16,7 @@ build_flags =
 	-DVECT_TAB_OFFSET=0x4000U
 board_build.ldscript = variants/R9M_ldscript.ld
 board_build.flash_offset = 0x4000
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 upload_flags =
 	BOOTLOADER=bootloader/r9m_bootloader.bin
 	VECT_OFFSET=0x4000
@@ -56,7 +56,7 @@ board_build.flash_offset = 0x8000
 upload_flags =
 	BOOTLOADER=bootloader/r9m_lite_pro_bootloader.bin
 	VECT_OFFSET=0x8000
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 
 
 # ********************************
@@ -73,7 +73,7 @@ build_flags =
 	-include target/Frsky_RX_R9M.h
 	-D HSE_VALUE=24000000U
 	-DVECT_TAB_OFFSET=0x08008000U
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/r9mm_bootloader.bin
     VECT_OFFSET=0x8000
@@ -98,7 +98,7 @@ build_flags =
 	-D HSE_VALUE=12000000U
 	-DVECT_TAB_OFFSET=0x08008000U
 board_build.ldscript = variants/R9MM/R9MM_ldscript.ld
-src_filter = ${common_env_data.src_filter} -<ESP32*.*> -<ESP8266*.*> -<WS281B*.*> -<tx_*.cpp>
+build_src_filter = ${common_env_data.build_src_filter} -<ESP32*.*> -<ESP8266*.*> -<WS281B*.*> -<tx_*.cpp>
 
 [env:Frsky_RX_R9SLIMPLUS_via_STLINK]
 extends = env:Frsky_RX_R9SLIMPLUS_via_BetaflightPassthrough
@@ -118,7 +118,7 @@ build_flags =
 	-D HSE_VALUE=12000000U
 	-DVECT_TAB_OFFSET=0x8000U
 	-D USE_DIVERSITY=1
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 
 [env:Frsky_RX_R9SLIMPLUS_OTA_via_STLINK]
 extends = env:Frsky_RX_R9SLIMPLUS_via_BetaflightPassthrough
@@ -143,7 +143,7 @@ build_flags =
 	-DVECT_TAB_OFFSET=0x08008000U
 	-D HSI_VALUE=16000000
 	-Wl,--defsym=FLASH_OFFSET=0x8000
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 upload_flags =
 	BOOTLOADER=bootloader/r9mx_bootloader.bin
 	VECT_OFFSET=0x8000
@@ -168,4 +168,4 @@ build_flags =
 	-D HSE_VALUE=12000000U
 	-DVECT_TAB_OFFSET=0x8000U
 board_build.ldscript = variants/R9MM/R9MM_ldscript.ld
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
