+++ src/targets/happymodel_2400.ini
@@ -12,7 +12,7 @@ build_flags =
 	-include target/HappyModel_ES24TX_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:HappyModel_ES24TX_2400_TX_via_WIFI]
 extends = env:HappyModel_ES24TX_2400_TX_via_UART
@@ -26,7 +26,7 @@ build_flags =
 	-include target/HappyModel_ES24TX_Pro_Series_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:HappyModel_ES24TX_Pro_Series_2400_TX_via_WIFI]
 extends = env:HappyModel_ES24TX_Pro_Series_2400_TX_via_UART
