+++ src/targets/happymodel_900.ini
@@ -15,7 +15,7 @@ build_flags =
 	-D VECT_TAB_OFFSET=0x4000U
 board_build.ldscript = variants/R9M_ldscript.ld
 board_build.flash_offset = 0x4000
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 upload_flags =
 	BOOTLOADER=bootloader/r9m_bootloader.bin
 	VECT_OFFSET=0x4000
@@ -33,7 +33,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_tx}
 	-include target/HappyModel_TX_ES900TX.h
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:HappyModel_TX_ES900TX_via_WIFI]
 extends = env:HappyModel_TX_ES900TX_via_UART
@@ -53,7 +53,7 @@ build_flags =
 	-include target/Frsky_RX_R9M.h
 	-D HSE_VALUE=24000000U
 	-DVECT_TAB_OFFSET=0x08008000U
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/r9mm_bootloader.bin
     VECT_OFFSET=0x8000
