+++ src/targets/iFlight_2400.ini
@@ -12,7 +12,7 @@ build_flags =
 	-include target/iFlight_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:iFlight_2400_TX_via_WIFI]
 extends = env:iFlight_2400_TX_via_UART
@@ -28,7 +28,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/iFlight_2400_RX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:iFlight_2400_RX_via_BetaflightPassthrough]
 extends = env:iFlight_2400_RX_via_UART
