+++ src/targets/iFlight_900.ini
@@ -10,7 +10,7 @@ build_flags =
 	${common_env_data.build_flags_tx}
 	-include target/iFlight_900_TX.h
 upload_speed = 460800
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 
 [env:iFlight_900_TX_via_WIFI]
 extends = env:iFlight_900_TX_via_UART
