+++ src/targets/imrc.ini
@@ -15,7 +15,7 @@ build_flags =
  	-D HSE_VALUE=32000000U
 	-DVECT_TAB_OFFSET=0x08004000U
 	-Wl,--defsym=FLASH_APP_OFFSET=0x4000
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 upload_flags =
 	BOOTLOADER=bootloader/ghost/ghost_tx_bootloader.bin
 	VECT_OFFSET=0x4000
@@ -43,7 +43,7 @@ build_flags =
  	-D HSE_VALUE=32000000U
 	-DVECT_TAB_OFFSET=0x08004000U
 	-Wl,--defsym=FLASH_APP_OFFSET=0x4000
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 lib_deps =
 lib_ignore = Servo
 upload_flags =
