+++ src/targets/namimnorc_2400.ini
@@ -14,7 +14,7 @@ build_flags =
 	-D VECT_TAB_OFFSET=0x4000U
 	-include target/NamimnoRC_FLASH_2400_TX.h
 board_build.ldscript = variants/NamimnoRC_Alpha.ld
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/namimnorc/tx/namimnorc_tx_bootloader.bin
     VECT_OFFSET=0x4000
@@ -31,7 +31,7 @@ build_flags =
 	-include target/NamimnoRC_FLASH_2400_OLED_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
 	${env_common_esp32.oled_lib_deps}
 
@@ -53,7 +53,7 @@ build_flags =
 	-D VECT_TAB_OFFSET=0x8000U
 	-include target/NamimnoRC_FLASH_2400_RX_STM32.h
 board_build.ldscript = variants/R9MM/R9MM_ldscript.ld
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/namimnorc/rx/flash_2400_bootloader.bin
     VECT_OFFSET=0x8000
@@ -68,7 +68,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/NamimnoRC_FLASH_2400_RX_ESP8285.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:NamimnoRC_FLASH_2400_ESP_RX_via_BetaflightPassthrough]
 extends = env:NamimnoRC_FLASH_2400_ESP_RX_via_UART
@@ -87,7 +87,7 @@ build_flags =
 	${common_env_data.build_flags_rx}
 	-include target/NamimnoRC_FLASH_2400_RX_ESP8285_PA.h
 	-D USE_DIVERSITY=1
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:NamimnoRC_FLASH_2400_ESP_RX_PA_via_BetaflightPassthrough]
 extends = env:NamimnoRC_FLASH_2400_ESP_RX_PA_via_UART
