+++ src/targets/namimnorc_900.ini
@@ -14,7 +14,7 @@ build_flags =
 	-D HSE_VALUE=12000000U
 	-D VECT_TAB_OFFSET=0x4000U
 board_build.ldscript = variants/NamimnoRC_Alpha.ld
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/namimnorc/tx/namimnorc_tx_bootloader.bin
     VECT_OFFSET=0x4000
@@ -32,7 +32,7 @@ build_flags =
 	-include target/NamimnoRC_VOYAGER_900_OLED_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
 	${env_common_esp32.oled_lib_deps}
 
@@ -54,7 +54,7 @@ build_flags =
 	-D VECT_TAB_OFFSET=0x8000U
 	-include target/NamimnoRC_VOYAGER_900_RX.h
 board_build.ldscript = variants/R9MM/R9MM_ldscript.ld
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 upload_flags =
     BOOTLOADER=bootloader/namimnorc/rx/voyager_900_bootloader.bin
     VECT_OFFSET=0x8000
@@ -70,7 +70,7 @@ build_flags =
 	${radio_900.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/NamimnoRC_VOYAGER_900_ESP_RX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:NamimnoRC_VOYAGER_900_ESP_RX_via_BetaflightPassthrough]
 extends = env:NamimnoRC_VOYAGER_900_ESP_RX_via_UART
