+++ src/targets/radiomaster_2400.ini
@@ -7,7 +7,7 @@ build_flags =
 	-include target/RadioMaster_Zorro_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 upload_speed = 460800
 lib_deps =
 	${env_common_esp32.lib_deps}
@@ -27,7 +27,7 @@ build_flags =
 	-include target/RadioMaster_TX16S_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 upload_speed = 460800
 lib_deps =
 	${env_common_esp32.lib_deps}
