+++ src/targets/siyi.ini
@@ -16,7 +16,7 @@ build_flags =
 	-DHSE_VALUE=16000000U
 	-DVECT_TAB_OFFSET=0x1000U
 	-Wl,--defsym=FLASH_APP_OFFSET=0x1000
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 upload_flags =
 	BOOTLOADER=bootloader/fm30_bootloader.bin
 	VECT_OFFSET=0x1000
@@ -42,7 +42,7 @@ build_flags =
 upload_flags =
 	BOOTLOADER=bootloader/fm30_mini_bootloader.bin
 	VECT_OFFSET=0x4000
-src_filter = ${env_common_stm32.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<tx_*.cpp>
 
 [env:FM30_RX_MINI_via_BetaflightPassthrough]
 extends = env:FM30_RX_MINI_via_STLINK
@@ -62,7 +62,7 @@ build_flags =
 upload_flags =
 	BOOTLOADER=bootloader/fm30_mini_rxtx_bootloader.bin
 	VECT_OFFSET=0x4000
-src_filter = ${env_common_stm32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_stm32.build_src_filter} -<rx_*.cpp>
 
 [env:FM30_RX_MINI_AS_TX_via_UART]
 extends = env:FM30_RX_MINI_AS_TX_via_STLINK
