+++ src/targets/vantac_2400.ini
@@ -12,7 +12,7 @@ build_flags =
 	-include target/Vantac_Lite_2400_TX.h
 	-D VTABLES_IN_FLASH=1
 	-O2
-src_filter = ${env_common_esp32.src_filter} -<rx_*.cpp>
+build_src_filter = ${env_common_esp32.build_src_filter} -<rx_*.cpp>
 lib_deps =
 	${env_common_esp32.lib_deps}
 
@@ -30,7 +30,7 @@ build_flags =
 	${radio_2400.build_flags}
 	${common_env_data.build_flags_rx}
 	-include target/Vantac_2400_RX.h
-src_filter = ${env_common_esp82xx.src_filter} -<tx_*.cpp>
+build_src_filter = ${env_common_esp82xx.build_src_filter} -<tx_*.cpp>
 
 [env:Vantac_2400_RX_via_BetaflightPassthrough]
 extends = env:Vantac_2400_RX_via_UART
